Fixed issues

+++ src/main/java/uk/gov/hmcts/juror/job/execution/jobs/dashboard/ams/data/BureauLettersToBePrintedDB.java
@@ -5,6 +5,8 @@ import lombok.Setter;
 import lombok.experimental.Accessors;
 import uk.gov.hmcts.juror.job.execution.database.DatabaseColumn;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Accessors(chain = true)
@@ -14,5 +16,5 @@ public class BureauLettersToBePrintedDB {
     @DatabaseColumn(name = "description", setter = "setDescription")
     public String description;
     @DatabaseColumn(name = "count", setter = "setCount")
-    public Long count;
+    public BigDecimal count;
 }
